openhcl_tdisp: move the no-op resource validator out of the mocks module
